Refactorización de la gestión de incidencias: se eliminan DTOs obsoletos, se actualizan entidades y repositorios para utilizar identificadores simples, y se mejora la lógica de creación y modificación de incidencias en el controlador.

+++ src/main/java/es/iesjandula/reaktor/issues_server/rest/IncidenciaController.java
@@ -17,20 +17,20 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import es.iesjandula.reaktor.base.security.models.DtoUsuarioExtended;
 import es.iesjandula.reaktor.base.utils.BaseConstants;
-import es.iesjandula.reaktor.issues_server.dto.CrearIncidenciaUsuarioDto;
 import es.iesjandula.reaktor.issues_server.dto.FiltroBusquedaDto;
 import es.iesjandula.reaktor.issues_server.dto.IncidenciaDto;
-import es.iesjandula.reaktor.issues_server.dto.ModificarIncidenciaResponsableDto;
 import es.iesjandula.reaktor.issues_server.dto.UsuarioCategoriaDto;
 import es.iesjandula.reaktor.issues_server.models.Categoria;
 import es.iesjandula.reaktor.issues_server.models.Incidencia;
-import es.iesjandula.reaktor.issues_server.models.ids.IncidenciaId;
+import es.iesjandula.reaktor.issues_server.models.Ubicacion;
 import es.iesjandula.reaktor.issues_server.repository.ICategoriaRepository;
+import es.iesjandula.reaktor.issues_server.repository.IUbicacionRepository;
 import es.iesjandula.reaktor.issues_server.repository.IIncidenciaRepository;
 import es.iesjandula.reaktor.issues_server.repository.IUsuarioCategoriaRepository;
 import es.iesjandula.reaktor.issues_server.utils.Constants;
@@ -77,6 +77,13 @@ import lombok.extern.slf4j.Slf4j;
 @RequestMapping(value = "/issues/incidencias")
 public class IncidenciaController
 {
+
+	/**
+	 * Repositorio de ubicaciones.
+	 */
+	@Autowired
+	private IUbicacionRepository ubicacionRepository;
+	
 	/**
 	 * Repositorio de incidencias.
 	 */
@@ -101,7 +108,10 @@ public class IncidenciaController
 	 * Este método recibe un objeto {@link CrearIncidenciaUsuarioDto} que contiene los datos de la incidencia a crear.
 	 * Realiza la validación de los datos y crea una nueva incidencia en la base de datos.
 	 * 
-	 * @param crearIncidenciaUsuarioDto El objeto {@link CrearIncidenciaUsuarioDto} que contiene los datos de la incidencia a crear.
+	 * @param usuario El usuario que crea la incidencia.
+	 * @param ubicacion La ubicación de la incidencia.
+	 * @param problema La descripción del problema de la incidencia.
+	 * @param nombreCategoria El nombre de la categoría de la incidencia.
 	 * @return Un objeto {@link ResponseEntity} que puede contener:
 	 *         <ul>
 	 *         <li>Un código de estado 200 (OK) si la incidencia se crea correctamente.</li>
@@ -111,15 +121,31 @@ public class IncidenciaController
 	 */
     @PreAuthorize("hasRole('" + BaseConstants.ROLE_PROFESOR + "')")
     @PostMapping("/")
-    public ResponseEntity<?> crearIncidencia(@AuthenticationPrincipal DtoUsuarioExtended usuario, @RequestBody CrearIncidenciaUsuarioDto crearIncidenciaUsuarioDto) 
+    public ResponseEntity<?> crearIncidencia(@AuthenticationPrincipal DtoUsuarioExtended usuario,
+											 @RequestHeader String nombreUbicacion,
+											 @RequestHeader String problema,
+											 @RequestHeader String nombreCategoria)
     {
         try 
         {
             // Validaciones previas sobre los datos de la incidencia
-			this.validarIncidencia(crearIncidenciaUsuarioDto);
+			this.validarCrearIncidencia(nombreUbicacion, problema, nombreCategoria);
+
+			// Buscamos la ubicación
+			Optional<Ubicacion> optionalUbicacion = this.ubicacionRepository.findById(nombreUbicacion);
+
+			// Validamos que la ubicación exista
+			if (!optionalUbicacion.isPresent())
+			{
+				log.error(Constants.ERR_INCIDENCIA_UBICACION_NO_ENCONTRADA_MESSAGE);
+				throw new IssuesServerError(Constants.ERR_INCIDENCIA_UBICACION_NO_ENCONTRADA_CODE, Constants.ERR_INCIDENCIA_UBICACION_NO_ENCONTRADA_MESSAGE);
+			}
+
+			// Obtenemos la ubicación
+			Ubicacion ubicacion = optionalUbicacion.get();
 
 			// Buscamos la categoría
-			Optional<Categoria> optionalCategoria = this.categoriaIncidenciaRepository.findById(crearIncidenciaUsuarioDto.getNombreCategoria()) ;
+			Optional<Categoria> optionalCategoria = this.categoriaIncidenciaRepository.findById(nombreCategoria) ;
 
 			// Validamos que la categoría exista
 			if (!optionalCategoria.isPresent())
@@ -132,7 +158,7 @@ public class IncidenciaController
 			Categoria categoria = optionalCategoria.get();
 
             // Buscamos los responsables de la categoría (el primero de la lista)
-            List<UsuarioCategoriaDto> responsablesCategoriaDto = this.usuarioCategoriaRepository.buscarResponsablesPorCategoria(crearIncidenciaUsuarioDto.getNombreCategoria());
+            List<UsuarioCategoriaDto> responsablesCategoriaDto = this.usuarioCategoriaRepository.buscarResponsablesPorCategoria(nombreCategoria);
 
 			// Validamos que los responsables existan
 			if (responsablesCategoriaDto.isEmpty())
@@ -146,12 +172,12 @@ public class IncidenciaController
 
             // Creamos la nueva incidencia
             Incidencia nuevaIncidencia = new Incidencia();
-			nuevaIncidencia.setUbicacion(crearIncidenciaUsuarioDto.getUbicacion());
+			nuevaIncidencia.setUbicacion(ubicacion);
 			nuevaIncidencia.setEmail(usuario.getEmail());
 			nuevaIncidencia.setFecha(LocalDateTime.now());
-			nuevaIncidencia.setDescripcion(crearIncidenciaUsuarioDto.getDescripcion());
+			nuevaIncidencia.setProblema(problema);
 			nuevaIncidencia.setEstado(Constants.ESTADO_PENDIENTE);
-			nuevaIncidencia.setComentario(null);
+			nuevaIncidencia.setSolucion(null);
 			nuevaIncidencia.setCategoria(categoria);
 			nuevaIncidencia.setEmailResponsable(responsableCategoriaDto.getEmailResponsable());
 
@@ -187,24 +213,29 @@ public class IncidenciaController
 	 * Este método valida los datos de la incidencia para asegurar que todos los campos requeridos estén presentes y sean válidos.
 	 * </p>
 	 *
-	 * @param crearIncidenciaUsuarioDto El objeto {@link CrearIncidenciaUsuarioDto} que contiene los datos de la incidencia a validar.
+	 * @param nombreUbicacion El nombre de la ubicación de la incidencia.
+	 * @param problema La descripción del problema de la incidencia.
+	 * @param nombreCategoria El nombre de la categoría de la incidencia.
 	 * @throws IssuesServerError si los datos de la incidencia no son válidos
 	 */
-	private void validarIncidencia(CrearIncidenciaUsuarioDto crearIncidenciaUsuarioDto) throws IssuesServerError
+	private void validarCrearIncidencia(String nombreUbicacion, String problema, String nombreCategoria) throws IssuesServerError
 	{
-		if (crearIncidenciaUsuarioDto.getUbicacion() == null || crearIncidenciaUsuarioDto.getUbicacion().isEmpty())
+		// Validamos el nombre de la ubicación
+		if (nombreUbicacion == null || nombreUbicacion.isEmpty())
 		{
 			log.error(Constants.ERR_INCIDENCIA_UBICACION_NO_INTRODUCIDA_MESSAGE);
 			throw new IssuesServerError(Constants.ERR_INCIDENCIA_UBICACION_NO_INTRODUCIDA_CODE, Constants.ERR_INCIDENCIA_UBICACION_NO_INTRODUCIDA_MESSAGE);
 		}
 
-		if (crearIncidenciaUsuarioDto.getDescripcion() == null || crearIncidenciaUsuarioDto.getDescripcion().isEmpty())
+		// Validamos el problema
+		if (problema == null || problema.isEmpty())
 		{
-			log.error(Constants.ERR_INCIDENCIA_DESCRIPCION_NO_INTRODUCIDA_MESSAGE);
-			throw new IssuesServerError(Constants.ERR_INCIDENCIA_DESCRIPCION_NO_INTRODUCIDA_CODE, Constants.ERR_INCIDENCIA_DESCRIPCION_NO_INTRODUCIDA_MESSAGE);
+			log.error(Constants.ERR_INCIDENCIA_PROBLEMA_NO_INTRODUCIDO_MESSAGE);
+			throw new IssuesServerError(Constants.ERR_INCIDENCIA_PROBLEMA_NO_INTRODUCIDO_CODE, Constants.ERR_INCIDENCIA_PROBLEMA_NO_INTRODUCIDO_MESSAGE);
 		}
 
-		if (crearIncidenciaUsuarioDto.getNombreCategoria() == null || crearIncidenciaUsuarioDto.getNombreCategoria().isEmpty())
+		// Validamos la categoría
+		if (nombreCategoria == null || nombreCategoria.isEmpty())
 		{
 			log.error(Constants.ERR_INCIDENCIA_CATEGORIA_NO_INTRODUCIDA_MESSAGE);
 			throw new IssuesServerError(Constants.ERR_INCIDENCIA_CATEGORIA_NO_INTRODUCIDA_CODE, Constants.ERR_INCIDENCIA_CATEGORIA_NO_INTRODUCIDA_MESSAGE);
@@ -214,10 +245,13 @@ public class IncidenciaController
 	/**
 	 * Actualiza una incidencia en el sistema por parte del responsable de la incidencia.
 	 * 
-	 * Este método recibe un objeto {@link ModificarIncidenciaResponsableDto} que contiene la información de la incidencia a
+	 * Este método recibe un objeto {@link IncidenciaDto} que contiene la información de la incidencia a
 	 * actualizar. Se realiza la validación de los datos y, si son válidos, se actualiza la incidencia en la base de datos.
 	 * 
-	 * @param modificarIncidenciaResponsableDto El objeto {@link ModificarIncidenciaResponsableDto} que contiene la información de la incidencia a actualizar.
+	 * @param id El ID de la incidencia.
+	 * @param estado El estado de la incidencia.
+	 * @param solucion La solución de la incidencia.
+	 * @param emailResponsable El email del responsable de la incidencia.
 	 * @return Un objeto {@link ResponseEntity} que puede contener:
 	 *         <ul>
 	 *         <li>Un código de estado 200 (OK) si la incidencia se actualiza correctamente.</li>
@@ -227,18 +261,18 @@ public class IncidenciaController
 	 */
     @PreAuthorize("hasRole('" + BaseConstants.ROLE_ADMINISTRADOR + "')")
     @PutMapping("/")
-    public ResponseEntity<?> modificarIncidencia(@RequestBody ModificarIncidenciaResponsableDto modificarIncidenciaResponsableDto)
+    public ResponseEntity<?> modificarIncidencia(@RequestHeader Long id,
+												 @RequestHeader String estado,
+												 @RequestHeader String solucion,
+												 @RequestHeader String emailResponsable)
     {
         try
         {
-			// Logueamos la incidencia a actualizar
-            log.info("Modificar incidencia DTO: {}", modificarIncidenciaResponsableDto);
-
-			// Creamos un identificador compuesto para la incidencia
-			IncidenciaId incidenciaId = new IncidenciaId(modificarIncidenciaResponsableDto.getUbicacion(), modificarIncidenciaResponsableDto.getEmailDocente(), modificarIncidenciaResponsableDto.getFecha());
+			// Validamos los datos de la incidencia
+			this.validarModificarIncidencia(id, estado, solucion, emailResponsable);
 
             // Buscamos la incidencia a actualizar
-            Optional<Incidencia> optionalIncidencia = this.incidenciaRepository.findById(incidenciaId);
+            Optional<Incidencia> optionalIncidencia = this.incidenciaRepository.findById(id);
 
 			// Validamos que la incidencia exista
             if (!optionalIncidencia.isPresent())
@@ -251,9 +285,9 @@ public class IncidenciaController
 			Incidencia incidencia = optionalIncidencia.get();
 
             // Actualizamos los campos que vienen en el dto
-			incidencia.setEstado(modificarIncidenciaResponsableDto.getEstado());
-			incidencia.setComentario(modificarIncidenciaResponsableDto.getComentario());
-			incidencia.setEmailResponsable(modificarIncidenciaResponsableDto.getEmailResponsable());
+			incidencia.setEstado(estado);
+			incidencia.setSolucion(solucion);
+			incidencia.setEmailResponsable(emailResponsable);
 
             // Guardamos la incidencia en la base de datos
             this.incidenciaRepository.saveAndFlush(incidencia);
@@ -279,6 +313,45 @@ public class IncidenciaController
     }
 
 	/**
+	 * Valida los datos de la incidencia.
+	 * @param id El ID de la incidencia.
+	 * @param estado El estado de la incidencia.
+	 * @param solucion La solución de la incidencia.
+	 * @param emailResponsable El email del responsable de la incidencia.
+	 * @throws IssuesServerError si los datos de la incidencia no son válidos
+	 */
+	private void validarModificarIncidencia(Long id, String estado, String solucion, String emailResponsable) throws IssuesServerError
+	{
+		// Realizamos las validaciones comunes
+		if (id == null || id <= 0)
+		{
+			log.error(Constants.ERR_INCIDENCIA_ID_NO_INTRODUCIDO_MESSAGE);
+			throw new IssuesServerError(Constants.ERR_INCIDENCIA_ID_NO_INTRODUCIDO_CODE, Constants.ERR_INCIDENCIA_ID_NO_INTRODUCIDO_MESSAGE);
+		}
+
+		// Validamos el email del responsable
+		if (emailResponsable == null || emailResponsable.isEmpty())
+		{
+			log.error(Constants.ERR_INCIDENCIA_EMAIL_RESPONSABLE_NO_INTRODUCIDO_MESSAGE);
+			throw new IssuesServerError(Constants.ERR_INCIDENCIA_EMAIL_RESPONSABLE_NO_INTRODUCIDO_CODE, Constants.ERR_INCIDENCIA_EMAIL_RESPONSABLE_NO_INTRODUCIDO_MESSAGE);
+		}
+
+		// Validamos el estado de la incidencia
+		if (estado == null || estado.isEmpty())
+		{
+			log.error(Constants.ERR_INCIDENCIA_ESTADO_NO_INTRODUCIDO_MESSAGE);
+			throw new IssuesServerError(Constants.ERR_INCIDENCIA_ESTADO_NO_INTRODUCIDO_CODE, Constants.ERR_INCIDENCIA_ESTADO_NO_INTRODUCIDO_MESSAGE);
+		}
+
+		// Validamos la solución de la incidencia
+		if (solucion == null || solucion.isEmpty())
+		{
+			log.error(Constants.ERR_INCIDENCIA_SOLUCION_NO_INTRODUCIDO_MESSAGE);
+			throw new IssuesServerError(Constants.ERR_INCIDENCIA_SOLUCION_NO_INTRODUCIDO_CODE, Constants.ERR_INCIDENCIA_SOLUCION_NO_INTRODUCIDO_MESSAGE);
+		}
+	}
+
+	/**
 	 * Lista los estados de las incidencias.
 	 * 
 	 * Este método devuelve una lista de los estados de las incidencias.
@@ -290,7 +363,7 @@ public class IncidenciaController
 	 *         </ul>
 	 */
 	@PreAuthorize("hasRole('" + BaseConstants.ROLE_ADMINISTRADOR + "')")
-	@GetMapping("/estados")
+	@GetMapping("/estados/")
 	public ResponseEntity<?> listadoEstadoIncidencias()
 	{
 	    try
@@ -320,7 +393,7 @@ public class IncidenciaController
 	 * 
 	 * Este método devuelve una lista de incidencias ordenadas por fecha.
 	 * 
-	 * @param pageable
+	 * @param pageable La página de incidencias a listar.
 	 * @return Un objeto {@link ResponseEntity} que puede contener:
 	 *         <ul>
 	 *         <li>Un código de estado 200 (OK) si la lista de incidencias se devuelve correctamente.</li>
@@ -334,7 +407,7 @@ public class IncidenciaController
 		try
 		{     
 			// Buscamos las incidencias ordenadas por fecha
-			Page<Incidencia> incidencias = this.incidenciaRepository.buscarIncidenciaOrdenadaFecha(pageable); 	    
+			Page<IncidenciaDto> incidencias = this.incidenciaRepository.buscarIncidenciaOrdenadaFecha(pageable); 	    
 
 			// Devolvemos la respuesta
 			return ResponseEntity.ok().body(incidencias);
@@ -358,7 +431,8 @@ public class IncidenciaController
 	 * Este método recibe un objeto {@link IncidenciaDto} que contiene los detalles de la incidencia a eliminar.
 	 * Se realiza la validación de los datos y, si son válidos, se elimina la incidencia en la base de datos.
 	 *
-	 * @param incidenciaDto El objeto {@link IncidenciaDto} que contiene los detalles de la incidencia a eliminar.
+	 * @param usuario El usuario que elimina la incidencia.
+	 * @param id El ID de la incidencia.
 	 * @return Un objeto {@link ResponseEntity} que puede contener:
 	 *         <ul>
 	 *         <li>Un código de estado 200 (OK) si la incidencia se elimina correctamente.</li>
@@ -366,17 +440,14 @@ public class IncidenciaController
 	 *         <li>Un código de estado 500 (Internal Server Error) si ocurre un error inesperado.</li>
 	 *         </ul>
 	 */
-	@PreAuthorize("hasRole('" + BaseConstants.ROLE_PROFESOR + "')")
+	@PreAuthorize("hasAnyRole('" + BaseConstants.ROLE_PROFESOR + "', '" + BaseConstants.ROLE_ADMINISTRADOR + "')")
 	@DeleteMapping("/")
-	public ResponseEntity<?> borraIncidencia(@RequestBody(required = true) IncidenciaDto incidenciaDTO)
+	public ResponseEntity<?> borrarIncidencia(@AuthenticationPrincipal DtoUsuarioExtended usuario, @RequestHeader Long id)
 	{
 		try
 		{
-			// Creamos un identificador compuesto para la incidencia
-			IncidenciaId incidenciaId = new IncidenciaId(incidenciaDTO.getUbicacion(), incidenciaDTO.getCorreoDocente(), incidenciaDTO.getFechaIncidencia());
-
 			// Buscamos la incidencia a eliminar
-			Optional<Incidencia> optionalIncidencia = this.incidenciaRepository.findById(incidenciaId);
+			Optional<Incidencia> optionalIncidencia = this.incidenciaRepository.findById(id);
 
 			// Validamos que la incidencia exista
 			if (!optionalIncidencia.isPresent())
@@ -387,6 +458,14 @@ public class IncidenciaController
 
 			// Obtenemos la incidencia
 			Incidencia incidencia = optionalIncidencia.get();
+
+			// Si el rol es de profesor, validamos que el usuario sea el que creó la incidencia
+			if (!usuario.getRoles().contains(BaseConstants.ROLE_ADMINISTRADOR) && !usuario.getEmail().equals(incidencia.getEmail()))
+			{
+				// Si llegamos aquí es porque el usuario no es el que creó la incidencia
+				log.error(Constants.ERR_INCIDENCIA_USUARIO_NO_PERMITIDO_MESSAGE);
+				throw new IssuesServerError(Constants.ERR_INCIDENCIA_USUARIO_NO_PERMITIDO_CODE, Constants.ERR_INCIDENCIA_USUARIO_NO_PERMITIDO_MESSAGE);
+			}
 
 			// Eliminamos la incidencia de la base de datos
 			this.incidenciaRepository.delete(incidencia);
